Refactoring

+++ src/GamificationProject/src/main/java/ch/heigvd/amt/gamification/dto/EndUserDTO.java
@@ -11,7 +11,7 @@ public class EndUserDTO {
     
     private String endUserNumber;
     private String apikey;
-    private int nbPoints;
+    private long nbPoints;
     private List<BadgeDTO> badges = new LinkedList<>();
 
     public String getEndUserNumber() {
@@ -34,11 +34,11 @@ public class EndUserDTO {
         return badges.size();
     }
 
-    public int getNbPoints() {
+    public long getNbPoints() {
         return nbPoints;
     }
 
-    public void setNbPoints(int nbPoints) {
+    public void setNbPoints(long nbPoints) {
         this.nbPoints = nbPoints;
     }
